fix: address review findings for RegisterIntEnum
- Fix broken doc comments: reorder RegisterEnum before RegisterIntEnum
  so each has its own doc block (2.1)
- Add alias collision panic to StringToIntEnumHookFunc (2.2)
- Add env-var and env-only integration tests for RegisterIntEnum (2.3)
- Validate empty/nil values map in RegisterIntEnum (2.7)
- Document int enum ordering by integer value (3.1)
- Document unsigned type exclusion (3.3)
- Fix enumHelpText int truncation on 32-bit: use int64 (3.2)

+++ internal/hooks/define.go
@@ -173,9 +173,9 @@ func DefineTimeDurationHookFunc() DefineHookFunc {
 // keyed by integer-typed levels. It returns the sorted value names and the
 // description with the addendum appended.
 func enumHelpText[L ~int | ~int8 | ~int16 | ~int32 | ~int64](levels map[L][]string, descr string) ([]string, string) {
-	keys := make([]int, 0, len(levels))
+	keys := make([]int64, 0, len(levels))
 	for k := range levels {
-		keys = append(keys, int(k))
+		keys = append(keys, int64(k))
 	}
 	slices.Sort(keys)
